pyproject.toml
[project]
name = "mine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-samples",
    "cairosvg>=2.7.1",
    "chess>=1.11.2",
    "fastapi[all]>=0.115.12",
    "httpx>=0.28.1",
    "jsonrpcclient>=4.0.3",
    "minio>=7.2.15",
    "pydantic-ai>=0.4.2",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=6.0.0",
]

[tool.uv.sources]
a2a-samples = { git = "https://github.com/google/A2A", subdirectory = "samples/python" }

